Extinction of Embedding Tissue boundary condition (DESIGN SURF EMBEDDING TISSUE CONDITIONS) (which actually only was a linear spring applied in every dof direction) and replacement by a more general spring-dashpot condition reading
-----------------------------DESIGN SURF SPRING DASHPOT CONDITIONS
DSURF 1
E 1 - spring_stiff 0.0 spring_offset 0.0 dashpot_visc 0.0 direction_all

Thus, to everyone who was using the embedding tissue, just replace it by this condition with your respective spring stiffness and zero dashpot viscosity, and you'll get the same thing.
Up to now, the displ and velo proportional reaction forces are applied in each dof direction (specified by direction_all). As a next step for the future, I'm planning to specify the alternative direction_surfnormal option to only take into account the normal movement of the surface dofs...

Additionally, it has left the PATSPEC section such that this one does not have to be tured on to apply the condition. I located it under drt_constraint so far - if anyone has a better place for it, go on.

And, the one and only testcase having made use of the embedding tissue was changed to the new condition.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@18927 060e22d8-9d1c-0410-868a-e3a950d63d95