globally switch to Trilinos Q3/2013.
We all have to thank Martin who did a great job to keep
BACI compiling for the continous nightly tests when someone
forgot to add all missing namespace expressions such as std::
Since this is a lot of work we now move to Q3/2013. The new
version is well-tested and stable. So, there is absolutely
no need to keep the experimental status.

There's currently no official "experimental" version.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@19034 060e22d8-9d1c-0410-868a-e3a950d63d95