ATTENTION NEW FRAMEWORK FOR STORING MATERIAL PARAMETERS
How it used to be:
Each material had a class to store its parameters which is derived from the base class
MAT::PAR::Parameter. Each material had a number of constitutive parameters which were stored in the derived class.
During the Evaluate call the material could access it's constitutive parameters through it's mat::par object directly via its name.

What was the problem?
For optimization and UQ we often want elements which have the same material
but have different material parameters (i.e. regional varying stiffness).


How it is now: (so far only changed in AAAneohook, other materials will follow)
In order to accommodate element specific material parameters,
all constitutive parameters are now stored in the
std::vector matparams_ in the class MAT::PAR.
The position in the vector is given by an emum in the respective derived class.

Each parameter is stored as Teuchos::RCP<Epetra_Vector> (within the std<vector>)
 which in the normal standard forward problem has the length 1!!
However, during optimization and UQ this Vector can be replaced by
a Epetra_Vector which has  element collumnmap layout,
storing a different value for each element.

Since the constitutive parameters are now potentially different for each
element the element ID is needed in each evaluation of the material law,
hence it is now passed as an argument in all evaluate functions.

Until now some material laws already the element ID for visualization.
However, it was "passed" to the evaluate function by storing it in the
parameter list before each call. This is very  inefficient, hence the change.

I ran ctest before this commit, but since not every material is tested please
make sure that your materials still work. Also I will  adopt the Elasthyper material
in the near future so please make sure that there is a testcase for each of your materials

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@19096 060e22d8-9d1c-0410-868a-e3a950d63d95