Reduce number of "minimal tests" from 36 to 24 without notable change of the code coverage of the minimal tests.
Note: if your test have been removed from the set of minimal tests it
is still tested during the nightly tests. So nothing is lost.

The idea of the minimal test set is to quickly (i.e. test in less than minute)
whether the most important functionality in BACI is working. In order to maximize
the code coverage of the minimal tests given above time constraint it is absolutely
contra-productive to test the same code with more than one tests.
Karl did a great job in finding a minimal set of tests with maximal code coverage.
Make sure that new minimal tests are as "orthogonal" as possible compared to the others.

Btw: Please also check your nightly tests and try to reduce the runtime.
We often test the same test on 1 and 2 processors. Most often it is probably
sufficient to keep only the 2 processor version.

Thanks in advance!

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@19388 060e22d8-9d1c-0410-868a-e3a950d63d95