removed outdated define flag FLUIDSPLITAMG:
It was needed to fill the condensed fluid interface velocity DOFs with 1.0 on the main diagonal in order to ensure a constant block size for the AMG aggregation. Furthermore, it is helpful to avoid changes in the range map of the matrix. Since we always want to do so, there is need for the define flag.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@19485 060e22d8-9d1c-0410-868a-e3a950d63d95