Adding additional IO flag "MATLAB_STATE_OUTPUT" into SCALAR TRANSPORT DYNAMIC
This flag will be checked at the end of the function SCATRA::ScaTraTimIntImpl::Output 
and will write the current solution into the ASCII file "Result_Step.m" via the function LINALG::PrintVectorInMatlabFormat
Be aware that in the current implementation only the !!last!! converged step is present in this file.
I will add soon Matlab functions, which can read this format.

The purpose of this output is a more convenient computation of the heart coordinate and heart fiber system via
the scripts heart_coords.py, lift_2d.py and lift.py.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@19535 060e22d8-9d1c-0410-868a-e3a950d63d95