Adding safety check for calling the boundary condition  in DESIGN SURF VOLUMETRIC FLOW CONDITIONS.
If this condition is set and the problemtype is still Fluid nothing will happen, since this condition
is only evaluated in Fluid_RedModels. Therefore if the baci detects this condition and the problemtype
Fluid_RedModels is not set, the following error message will raise

ERROR: Given Volumetric Womersly infow condition only works with Problemtyp Fluid_RedModels. 
 --> If you want to use this conditions change Problemtyp. 
 --> If you don't want to use this condition comment the respective bc section.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@19686 060e22d8-9d1c-0410-868a-e3a950d63d95