changed Gauss integration rule for So3_Plast elements ===================================================== - changed Gauss integration from drt_utils_gauuspoint.H to drt_utils_integration.H - this ensures, that the same integration rule is used in So3_Plast as well as in thermo_ele_impl - needed for upcoming TSI
Be aware, that the Gauss points for the hex8 element are ordered differently in the two integration rules. This may lead to inconsistencies when GP data are transferred between different fields that use different GP numbering. 
The structural hex8 element has its own Gauss point locations coded, which are identical to the drt_utils_integration.H integration also used in thermo_ele_impl.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@19705 060e22d8-9d1c-0410-868a-e3a950d63d95