A nonlinear Solver Package: ===========================
In the new source code folder 'solver_nonlin', a set of nonlinear solution methods is provided. The namespace is NLNSOL. There are two main groups of classes:
- Nonlinear Operators: Here, the solution algorithm is implemented in an abstract fashion.
- Nonlinear Problem: This is the interface between the nonlinear solver and the time integrator in baci (e.g. for Evaluate()-calls)

Up to now, four NlnOperator-objects are implemented:
- NlnOperatorNGmres: A nonlinear GMRES algorithm (requires another NlnOperator as nonlinear preconditioner)
- NlnOperatorFas: A nonlinear algebraic multigrid algorithm using the full approximation scheme (AMG hierarchy is based on MueLu, requires a hierarchy of multigrid levels, requires NlnOperators as nonlinear level smoothers)
- NlnOperatorQuasiNewton: A Quasi-Newton implementation with fixed Jacobian
- NlnOperatorNonlinCG: A nonlinear CG algorithm

For multigrid, we additionally need a coarse level representation of the nonlinear problem interface: NlnProblemCoarseLevel as well as FAS specific stuff in the NLNSOL::FAS namespace.

A set of line search methods is derived from NLNSOL::LineSearchBase. Available methods are:
- LineSearchFullStep: just take the full step
- LineSearchPolynomial: Build a second order polynomial model
- LineSearchBacktracking: try a reduction of the step size step by step (This has a notion of 'brute force' and is, thus, not recommended.)

Configuration of the nonlinear solver is done via an XML-File. NLNSOL::Utils provides routines to convert the XML-File to a (nested) Teuchos::ParamterList.

An exemplary connection to the structural time integrator is done. 

No test, yet. No XML-Files, yet. No guarantee for convergence, yet. No great performance, yet.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@19710 060e22d8-9d1c-0410-868a-e3a950d63d95