Adapted and renamed the "BC-Free Boundary Condition" to "Slip Supplemental Curved Boundary Condition".
The slip condition consists of a Dirichlet condition, i.e. the normal
velocities at the boundary are zero, and a Neumann condition, i.e. the
tractions in tangential direction are zero.

For non-curved boundaries, a slip condition can easily be realized in
BACI by just setting the normal coordinate of velocity to zero (by use
of the locsys functionality). The condition that the tangential
traction is zero is automatically (indirectly) fullfilled by BACI
standard behavior of setting "do-nothing" conditions, i.e. traction
equals zero at boundaries where no other conditions have been
specified.

For curved boundaries the situation is different. Even though the
tangential tractions have been set to zero, when integrating the
normal boundary tractions at the nodes, the resulting vector at the
node has a tangential component, which does not make sense in the
continuous case, but is simply a discretization error. A nice
explanation of this issue is given in Behr M., 2003, "On the
Application of Slip Boundary Condition on Curved Boundaries".

The result of this tangential traction component at the boundary is
that spurious velocities arise. In the past, I used a solution to this
recommended in the paper by Marek Behr as cited above. This solution
successfully counteracts the spurious velocities, but also alters the
real exisiting traction at the boundary. Due to the fact that within
my research topic I am specifically interested in this traction, I
implemented another solution recommended by Coppola-Owen H. & Codina
R., 2011, "A free surface finite element model for low Froude number
mould filling problems on fixed meshes". This solution finally
provides the proper tractions at the boundary while counteracting the
spurious velocities coming from the discretization error.

Thanks to Ursula and also Martin for many discussions about this
issue.



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@19731 060e22d8-9d1c-0410-868a-e3a950d63d95