o A new problem type "TWO PHASE FLOW" was created. This problem type utilizes a coupling between the level set solver and the standard fluid solver. In due time it is expected to transfer the capabilities of the combust module here.
o For the time being a CSF (Continuum Surface Force) model is implemented. This algorithm uses a smoothing function to "smear" the discontinuity (density and viscosity) between two fluids. This is done within an interface, which thickness is user provided. Within this interface a surface tension force is also applied, thus also "smearing" the surface tension force.
  - Further implementation in this model will be the usage of particles to provide better mass conservation properties.

o To be able to calculate the surface tension, the nodal values of the gradient and the curvature of the level set is needed. These functions are implemented in the classes SCATRA::ScaTraTimIntImpl and SCATRA::LevelSetAlgorithm respectively. 

o In FluidEleCalc::GetMaterialParams() a blending function for an input of a material list is implemented. At the moment only 2 Newtonian fluids can be blended.

o To test against experimental results for a rising bubble, a function to calculate the mass center (utilizing the blending function) was also implemented in SCATRA::LevelSetAlgorithm.

o The convergence check for the scalar transport equation, previously residing in LOMA, was moved to ScaTraTimIntImpl, providing access to other instances.

o 5 test cases was added for the two phase flow problem utilizing the CSF model.

Big thanks to Ursula for all the help and guidance!

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@19744 060e22d8-9d1c-0410-868a-e3a950d63d95