Fix problems when the second block has more then one dof per node.
This fix also requires changes in Trilinos (in Xpetra and MueLu).
Some preprocessor guards are added in Baci to skip the lines related with this
change ensuring that baci still compiles with the standard Trililos version

As soon as the standard trilinos version includes these changes, we will
delete the guards.

Thanks to Tobias for the help provided with Trilinos issues.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@19758 060e22d8-9d1c-0410-868a-e3a950d63d95