Added functionality to consider uncertain geometries (wall thickness ) in UQ framework (Input flag in MLMC section RANDOMGEOMETRY =ON)
Random geometries can  be generated by moving nodes normal to a surface, the distance of
the shift is determined by a random field. In order to obtain surface normals a mortar interface
is created for the sole purpose of nodal normal evaluation via a specialized setup routine of
the mortar adapter.

To create aneurysm models with uncertain wall thickness with multiple elements across the
thickness of the wall, an ALE discretization is created to move the nodes within the wall. 
This way only the translation of the nodes on the surface is given by the field and the
position of the inner nodes is adjusted automatically. The nodes that comprise the uncertain 
surface are defined via the new UNCERTAIN SURFACE CONDITION.


specifically:
added:
new class mc_var_thickness_manager to handle the geometry adjustments during
runtime using an ALE discretization which is created by cloning the structure
discretization. This class also handles the setup for the random field which describes the wall thickness.
new condition uncertain surface

modified:
Class adapter_coupling_mortar now has a specialized setup routine for UQ purposes, i.e.,
 setup an interface based on the nodes in the uncertain surface to compute
averaged nodal normals
Class drt_node has another method to change the reference position
averaged nodal normals can now be extracted from the mortar_interface

Class random_field_fourier:
Bugfix: Periodicity of fields generated with Fourier series is now an input parameter and
no longer determined by the dimensions of the discretization

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@19767 060e22d8-9d1c-0410-868a-e3a950d63d95