P-ADAPTIVITY
This commit enables p-adaptivity for acoustic HDG elements. Both acoustic elements and all corresponding time integrators are converted such that the internal field variables are no longer stored as global vector but each element stores its values by itself. This code design seems appealing for elementwise discontinuous variables. The trace field is still stored in a global vector. In case of p-adaptivity (flag P_ADAPTIVITY to Yes), an error estimator is evaluated (by use of the superconvergent post processing) and a new degree is calculated. Then, the internal field is projected onto the desired polynomial basis and stored in the element variables (that's why its usefull to store the variables on element level). We do not have to project the trace field to the new polynomial basis since no time derivative of the trace variable occurs in the underlying PDEs. 

Nonetheless, there are some restrictions which will be adressed as soon as possible: p-adaptivity does not yet work for dirk time integration schemes, and does not yet work in parallel. Also, I will create a test case for p-adaptivity in near future and adpat the restart functions.

- introduced the internal field variables in the acoustic elements
- removed the global vectors holding the internal field in the time integrators
- wrote function which initializes the elemental field variables depending on the polynomial basis (which does not change if no p-adaptivity is performed)
- introduced function UpdatePolyAndState() which updates the degree of the elements, rebuilds the faces and the dofmaps
- had to make the function AssignDegreesOfFreedom public
- removed specialized UpdateInteriorVariablesAndAssemebleRHS() and AssembleMatAndRHS() funciton for BDF time integration, since they are no longer needed for the field storage concept

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@19773 060e22d8-9d1c-0410-868a-e3a950d63d95