Adaptive time step for pure fluid based on the CFL number.
Two new fluid parameters are introduced:

- CFL_NUMBER
- FREEZE_ADAPTIVE_DT_AT

The first parameter specifies the CFL number. Reasonable values are somewhere in the region of 0.1-1.5 for most applications.
The second parameter keeps the time step constant after the specified time step value. This is necessary for consistent sampling of turbulence statistics data.
If the velocity field is zero in the beginning of the simulation, the value specified in the input file is taken as first time step and a warning is printed to screen. 

The time step is calculated with the following formula:

dta_ += C*(CFL*h/|u|-dtp_),

where dta_ and dtp_ are the new and old time step, h is the element length calculated with the stream-length method, |u| is the norm of the velocity and C is a constant increasing robustness and avoiding oscillations (0.8 hard-coded at the moment). 
The minimum value of all elements is taken as final time step.
Steady-state is assumed, i.e. the time step is not repeated if the condition would require a smaller step in the new time step. This might cause problems in highly instationary flows.

Please report your experience with this method and suggestions for improvements to me.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@19825 060e22d8-9d1c-0410-868a-e3a950d63d95