A framework for cutting a mesh with a level set function has been produced. As this procedure shares functionalities with the already existing cut framework, inheritance is utilized to provide access to functions which both frameworks use.
o Inheritance and accessing of the different classes is shown in the illustration below.


                 |FluidWizardLevelSet| ------accesses----->  |CutWizardLevelSet|   ------accesses------->   |LevelSetIntersection|
                /                                           /                                              /
base class:    /                              base class:  /                            base class:       /
---------------                               -------------                         ----------------------
| FluidWizard |                               | CutWizard |                         | ParentIntersection |
---------------                               -------------                         ----------------------
               \                                           \                                              \
                \                                           \                                              \
                 |FluidWizardMesh|  ---------accesses----->    |CutWizardMesh|     ------accesses------->   |MeshIntersection|     
                 
                 
                 
o In files where the FluidWizardMesh is not specifically required, the base class FluidWizard is used. This gives the possibility to use these functions with the FluidWizardLevelSet as well, for instance the Dofset handling.
                 
o The cut frameworks are specified in the respective classes, including the addition of elements to the cut mesh. Almost all of the getters are provided in the base class (except for CutSides, as this does not exist in the level set cut framework), also the assigning of DoF's are done in the base classes.                  
                 
o As this new framework required the old FluidWizard to be renamed into FluidWizardMesh and the old CutWizard into CutWizardMesh, a few files needed to be accessed to change the current naming convention.

o The level set cut requires the input of a mesh and a node column vector of the level set values at the nodes of the given mesh (For two phase flows this means that the scatra discretization and fluid discretization need to coincide).

Big thanks to Benedikt for the help and guidance! I would also like to direct a thank you to Andy, Sudhakar, Raffaela and Shardan for helping me out as well!

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@19870 060e22d8-9d1c-0410-868a-e3a950d63d95