deleted functions PrepareTimeandStep() and UpdateStepandTime() and replaced them by IncrementTimeAndStep()
previously the TurbulentFlowAlgorithm used PrepareTimeandStep(), which then called UpdateStepandTime(), to update the time and the step counter. These functions have the same functionality as IncrementTimeAndStep().
To avoid confusion in future we only use IncrementTimeAndStep() to increment the time and the step counter. The routine PrepareTimeStep() with completely different functionality remains unchanged.



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@19883 060e22d8-9d1c-0410-868a-e3a950d63d95