New memory management of LINALG::SerialDense{Matrix,Vector} ===========================================================
Previously, Epetra_SerialDenseMatrix (and thus, LINALG::SerialDenseMatrix) re-allocated memory whenever a call to Shape() or Reshape() was issued. This leads to frequent memory re-allocation which is expensive. If the matrix sizes stay the same, the memory allocation can be avoided by checking the sizes before allocating. This does not only lead to many unnecessary lines in codes that uses SerialDenseMatrix, but it also does not help in situations where the matrix sizes frequently change between a few values (e.g., p-adaptivity for acoustic HDG where we need to resize local data structures when the degree from one element to the next changes).

The new implementation applies a strategy similar to what is done in std::vector::resize(n). We remember the amount of memory allocated and if the new matrix size fits with the allocated memory, we re-use it. If the matrix size increases, we of course need to still allocate (so be careful when slowly increasing the matrix/vector sizes with Reshape()/Resize() because data is copied every time). If you want to keep tight memory upon Shape() without unnecessary memory (e.g. large matrices), set the size to zero first and then the desired value. Beware that LINALG::SerialDenseMatrix still inherits from Epetra_SerialDenseMatrix and the new Shape() method is not virtual: If you pass a linalg matrix to a function with Epetra_SerialDenseMatrix argument and call Shape() there, the base class method that unconditionally allocates memory is used. This does not break code but does not save anything.

The new matrices are used in the HDG polynomials, HDG shape functions, and the BoundaryGPToParentGP functions. For the latter one to work, I changed matrix types from Epetra_SerialDenseMatrix to LINALG::SerialDenseMatrix in fluid_ele_boundary_calc_poro.cpp and wall1_line_evaluate.cpp.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@19893 060e22d8-9d1c-0410-868a-e3a950d63d95