Extended the current capabilities of the two phase flow utilizing standard FEM and level set with smoothing function.
o Particles can be added in these simulations for better conservation properties of the level set.
o Output capability is also added for two phase flow calculations with particles.
o A linear interpolation is used to connect level set one-step-theta simulations with fluid gen-alpha simulations.
o To increase accuracy in the interface layer, a parameter ENHANCED_GAUSSRULE is introduced. If set to Yes, a gauss rule suitable for sine and cosine is utilized in the interface layer. The implementation can be seen in fluid_ele_calc.cpp. If there is need to test other gauss rules, changes should be made there.

As always, thanks to Ursula for help and input!

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@19958 060e22d8-9d1c-0410-868a-e3a950d63d95