Clean-up Moved UQ framework out of STR namespace Removed several old and unused input flags from mlmc section Removed some unused functions in mlmc.cpp Added new problem type UQ
Added choice of UQ strategy
As of now, the user can choose between

mc_plain:
i.e. standard Mc

mc_parametercontinuation:
instead of solving the full nonlinear problem for each sample
keep the load level fixed and gradually adapt the material properties
(Note, this strategy does not yet work for stochastic wall thickness)

mc_scaledthickness:
instead of solving the full nonlinear problem the von Mises stress
is computed based on the values form an initial run with constant thickness and simply scaled
by the local wall thickness, that is obtained from a random field.
(still experimental)

Bugfix when considering stochastic wall thickness in Combination with MULF
Reset of MULF history variables after adjusting the mesh via an ALE step

Started extension of UQ framework to red airways

Added basic framework to perform UQ with red airways framework
Currently it is not yet possible to perform proper UQ with red airways yet
I've just added a loop around the red airways time integration allowing multiple
calls to (up to now )the same forward problem. It also does not work with
copyDatFile type nested parallelism.


Bugfix/Enhancement in patspec section

Up to now the maximum ILT thickness was computed as maximum distance of any node to the
luminal surface of the ILT, which was identified by either orthopressure or fsi interface.
This obviously included wall elements. To account for that an assumed wall thickness of
1mm was subtracted from the max ILT thickness. This leads to problems when considering
a spatially varying wall thickness, in particular for very thin walls
As a remedy a new method to compute the max ILT thickness accurately was introduced.
This requires the knowledge of the luminal ILT surface and the outer ILT surface, which can be
provided through the new AAA surface condition.
The accurate max ILT thickness calculation can be switched on by setting
the flag CALC_ACCURATE_MAX_ILT_THICK to yes.
The old variant is still kept to allow for an consistent re-evaulation of our AAA database.

Added Testcase for stochastic wall thickness and Fourier series representation of random field

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@20036 060e22d8-9d1c-0410-868a-e3a950d63d95