Inverse Analysis of Material Elast_Coupexppol with the SEF \Psi = a \exp \left( b(I_1-3) - (2b+c)ln(J) + c(J-1) \right) -a
For a better Inverse Analysis of the Material Parameters of Elast_Coupexppol the Inverse Analysis of the Parameters b and c is done with ln(b) and ln(c) instead of b and c. 
Prints on Screen in Inverse Analysis are recalculated, so one sees the values for the real parameters b and c. Also Input-Line should include real Parameters b and c.
Just Inverse Analysis uses ln(b) and ln(c), normal calculations use b and c.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@20039 060e22d8-9d1c-0410-868a-e3a950d63d95