added total variation (TV) regularization to the inverse analysis
the implementation is based on a graph representation of the data and an equivalent TV-norm based on finite difference gradient approximations for the graph data.

the graph (actually the weighted graph) is handed as intput to the regularization. It can be arbitrary but must fit the optimization parameter map 
of course. Thus it can be constructed by the MatParManager classes since they take care about the paramter vector layout anyways.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@20119 060e22d8-9d1c-0410-868a-e3a950d63d95