NEW ALE
Since we found several inconsistencies and bugs in the current ALE implementation and had nothing else to do, we decided to cleanup the ALE framework.

We reorganised the ALE framework and introduced an adapter layer. To minimize transitional work that needs to be done to move to this new ALE framework, the new implementation currently resides in the folder drt_ale_new/. Every ALE user is required to use the adapter in the future as interface to the ALE field. We have started to prepare problem-type specific adapter classes that derive from the adapter wrapper. Just like in the structure. 

The users of the different problem types should implement these classes and move all the problem specific stuff, that used to be in ale.cpp or ale_lin.cpp or ..., into the respective adapters. The ALE class itself must not be touched. This should happen until november 1st. After that the old implementation will be removed.

The old implementation is still in place, though it cannot be guaranteed that you will obtain the same results as before because we have fixed several bugs on the ale-element level.
Hence, it is possible that some test cases will fail.  Please check your test cases tomorrow and make sure that they are failing only because of slightly different mesh movement.

In the new ALE framework the following ALE types can be chosen:

solid 	 		based on structural solid assumption; (works with any material from the ELAST_HYPER toolbox as well)

springs 	 	analogy to a spring system where nodes are connected by lineal springs and additional torsional springs in the element corners 

laplace	 		old implementation, still looks very dodgy!Useful results cannot be guaranteed and hence the implementation currently contains a dserror and should not be used until someone has verified its validity

In the ALE DYNAMIC section the user can now additionally specify the number of Newton steps and tolerances to solve the ALE problem. 
If a quasi linear formulation is desired just use (also default parameters)
MAXITER  	1
DIVERCONT 	continue

This works only if the ALE field is used in a partitioned algorithm or as single field problem. In the monolithic case, the Newton of the coupled problem takes care of the nonlinear solution of the ALE field as well.

In the process of testing the new ALE framework, we found that our (not so new) SOLID3 element is not yet tested in any of the nightly tests. Hence, we have added a test case for the 
TET10 version of this element. 

Matthias and Jonas

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@20125 060e22d8-9d1c-0410-868a-e3a950d63d95