===| Introduce Base Class Field |===
A new Base Class for all types of Field is introduced. There are just methods, which should be available for all kinds of fields. With this commit
structure and poro derives from this base class.

- general Field Wrapper and PoroStructure Wrapper is introduced to exchange Structures and Poro in the fpsi_xfem algorithm (algorithm commit will come later)
- poro is modified to have the same methods as most of the other fields have
		- evaluate now sets up the whole system matrix and the rhs (no call of setupRHS() from outside)
		- moved updateporoiterinc() into preparetimestep()
		- renamed some methods
- modifications in fpsi

Thx to Georg, Karl, Martin, Raffaela, Benedikt, Anh-Tu for the discussion on the field adapter.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@20282 060e22d8-9d1c-0410-868a-e3a950d63d95