Fixed bug in calculation of decoupled elasthyper materials in Elasthyper-Toolbox ================================================================================
* changed computation of cmati in EvaluateIsotropicMod in elasthyper
* changes affect the following material:
elast_iso2pow (with d>=2)

Check your results if you use this material!

Thanks to Alex S. for the help to find the bug!


New Material IsoTestMaterial
============================

This material is not realistic, but contains all possible isochoric (first and second) derivations of invariants.
SEF:
\Psi = C1 (\overline{I}_{\boldsymbol{C}}-3) + 0.5 C1 (\overline{I}_{\boldsymbol{C}}-3)^2
     + C2 (\overline{II}_{\boldsymbol{C}}-3)  + 0.5 C2 (\overline{II}_{\boldsymbol{C}}-3)^2
     + D (\overline{I}_{\boldsymbol{C}}-3) (\overline{II}_{\boldsymbol{C}}-3)
With this material and the material volsussmanbathe, it is possible to test all parts of the Elasthyper-Toolbox.


New Testcase elasthyper_toolbox_testmaterial
============================================

This testcase tests all parts of the stress-tensor and constitutive-tensor in the elasthyper-toolbox.
Therefore the new material elast_isotestmaterial and elast_volsussmannbathe are used.
Tests all possible distortion conditions.

Thanks to Andi N. for the testcase!
Thanks to Martin P. for the script to create the result discription!

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@20323 060e22d8-9d1c-0410-868a-e3a950d63d95