--------------------------------------- ad_fld_base_algorithm:
move the functionality to set an intial flow field and an initial inflow field into own routines,
as for Fluid-LevelSet-coupling we have to call SetInitialFlowField independent of the Init-call
-> SetInitialFlowField
-> SetInitialInflowField

Changes in adapter_scatra_fluid_coupling_algorithm: for prb_fluid_xfem_ls:
- we first have to call the Fluid-Init.
- Second, using a levelset-field we create the new state-class.
- After that we can do SetInitialFlowField


---------------------------------------
xfluid:
- gmsh output for wedge/pyramid elements


xfluid with levelset:

- own CreateInitialState routine to create the state-class from the xfluid_level_coupling_algorithm
- fixed some calls for level-set
- ComputeErrorInterface get an RCP to boundarydis instead of reference
- SetLevelSetField()-routine to set phinp-vector in fluid from xfluid_level_coupling_algorithm
- get material in fluid_ele_calc_xfem as we will need it for two-phase flow
- reorganized the XFLUIDLEVELSET::Algorithm

---------------------------------------
special pressure symmetry condition for pure-Dirichlet Pseudo-2D prblems solved with Krylov (PRES_KRYLOV_2Dz = false(default)):
Using pressure-edge-based stabilization for pure Dirichlet-problems, the pressure layers in z-direction are completely decoupled.
In that case a classical Krylov is not sufficient to get a unique pressure solution.
In contrast, PSPG-stabilization enforces a hidden constraint, namely \int_\Omega (\grad p, \grad q) = 0, which ensures that the pressure gradient in z-direction has to be minimal in a least-squares sense.
Therefore, PSPG couples front- and backlayer of the pressure field in this case.
We enforce a similar constraint for pure pEOS stabilization:
\int_F \dq/dz dp/dz dF = 0 which minimizes the pressure gradient in z-direction. This condition is only reasonable for Pseudo2D simulation solved with Krylov where we expect symmetric solutions in z-direction.

-> added new testcase (tests Oseen problem type in combination with special pressure symmetry condition and Krylov projection) 


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@20358 060e22d8-9d1c-0410-868a-e3a950d63d95