------------------------------------------------- CHANGE in the integration of elements in the XFEM -------------------------------------------------
SO FAR: all elements which are handled via the cut-wizard, i.e. also elements which are not intersected by the interface, however, are contained in a bounding box around the boundary-discretization or elements which are not intersected, however, have negative level-set values at all nodes,
have been integrated by quite accurate XFEM-specific integration rules of high order.

In our fluidcode, however, the stabilized weak formulation on standard fluid elements is not integrated exactly because of efficiency reasons. As the polynomial order of, e.g., the SUPG term is higher than the used Gauss integration rule. This leads to integration errors for uncut fluid elements.

NOW: only elements which are really intersected by the interface are integrated via our XFEM-integration rules. Uncut elements (although handled in the cut-library and GEO::CUT::Elemenhandles are available) are integrated as classical fluid elements. This gives speedup in certain cases, however, changes some results.

-> This commit fixes the following level-set testcase xfluid_forward_facing_step_stationary_levelset_16x16.dat
and gives same results as before commit 20454. Now, the same elements are integrated via XFEM-integration rules as the corresponding testcases using a mesh-cut. The intermediate failing of this test was because of changes in the handling of elements via the CutWizardNEW.

-> CHANGING RESULTS in xff_beltrami tests and some xfluid_multibody_..., xfluid_triple_body_... testcases
	- the changes of the results due to the different numerical intergration is of about 1e-08 and is well studied
	- NOTE: Corresponding tests, e.g. xfluid_multibody_hex8_EOS_... do not give changing results, as no higher-order SUPG/PSPG stabilization are used there!
	
-> Now, all xfluid/xfluidfluid results should be independent of which or how many non-intersected elements are handled in the cut-wizard!!!

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@20461 060e22d8-9d1c-0410-868a-e3a950d63d95