Fixed bug in FindElementConditions which was introduced in commit 9951 by Uli.
For a given element, the routine returns all conditions for given conditionname that all nodes of this element have in common.
This bug occurred only when the nodes of one element have been conditioned by more than one condition in the .dat-file.
Fortunately, in most of the cases only one condition is set in the input file for the condition types for which FindElementConditions() is used.

Add additional comments in this function.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@20494 060e22d8-9d1c-0410-868a-e3a950d63d95