AC-FSI can now periodically repeat the fsi problem. It's done by reading from the output file all values of the fsi problem written one period before. Input looks like this: ----------------------------------------------------------------FS3I DYNAMIC/AC FSISTEPSPERSCATRASTEP           2 PERIODICITY                     0.03 PERIODSTOSTEADYSTATE            1 PERIODSTOFSIUPDATE              4
FSISTEPSPERSCATRASTEP: Subcycling of the fsi problem
PERIODICITY: Time of one period of the fsi problem
PERIODSTOSTEADYSTATE: number of periods after which the fsi problem is assumed to be in steady state
PERIODSTOFSIUPDATE: number of periodically repeated fsi cycles before it updates the period via standard calculations

Periodical repetition also works in combination with subcycling. Added another testcase to test this.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@20515 060e22d8-9d1c-0410-868a-e3a950d63d95